TE: add some fixes to the campaign full
