+++ tests/E2E/test/campaigns/common_scenarios/shop_parameters.js
@@ -85,5 +85,15 @@ module.exports = {
       await client.checkTextValue(OnBoarding.welcomeSteps.message_value, 'Here are the shipping methods available on your shop today.', 'equal', 2000);
       await client.scrollWaitForExistAndClick(OnBoarding.welcomeSteps.next_button);
     }
+  },
+  enablePrestashopDebugMode: function (Menu, Performance) {
+    if (!global.ps_mode_dev) {
+      scenario('Enable the debug mode', client => {
+        test('should go to "Performance" page', () => client.goToSubtabMenuPage(Menu.Configure.AdvancedParameters.advanced_parameters_menu, Menu.Configure.AdvancedParameters.performance_submenu));
+        test('should switch the "Debug mode" to "Yes"', () => client.waitForExistAndClick(Performance.enableDebugMode));
+        test('should click on "Save" button', () => client.waitForExistAndClick(Performance.save_button.replace('%I', 2)));
+        test('should verify the appearance of the green validation', () => client.checkTextValue(Performance.success_box, "Successful update."));
+      }, 'common_client');
+    }
   }
 };
