+++ tests/E2E/test/campaigns/full/04_attribute/2_create_attribute.js
@@ -98,11 +98,7 @@ scenario('Create "Attributes" in the Back Office', () => {
       return promise
         .then(() => client.isVisible(ProductList.pagination_products, 3000))
         .then(() => client.getProductsNumber(ProductList.pagination_products))
-        .then(() => {
-          if (global.ps_mode_dev) {
-            client.waitForExistAndClick(AddProductPage.symfony_toolbar)
-          }
-        });
+        .then(() => client.waitForSymfonyToolbar(AddProductPage, 2000))
     });
     test('should go to "Shop Parameters - Product Settings" page', () => {
       return promise
