+++ tests/E2E/test/campaigns/full/14_discount/2_check_discount_in_FO.js
@@ -74,11 +74,7 @@ scenario('Create "Catalog price rule"', () => {
     test('should go to "Products" page', () => client.goToSubtabMenuPage(Menu.Sell.Catalog.catalog_menu, Menu.Sell.Catalog.products_submenu));
     test('should close the "Symfony" toolbar', () => {
       return promise
-        .then(() => {
-          if (global.ps_mode_dev) {
-            client.waitForExistAndClick(AddProductPage.symfony_toolbar);
-          }
-        })
+        .then(() => client.waitForSymfonyToolbar(AddProductPage, 2000))
         .then(() => client.pause(1000));
     });
     test('should search for the created product', () => client.searchProductByName(productData["name"] + date_time));
