+++ tests/E2E/test/campaigns/high/02_product/04_create_virtual_product.js
@@ -29,7 +29,7 @@ scenario('Create virtual Product in the Back Office', client => {
     test('should upload the first product picture', () => client.uploadPicture('image_test.jpg', AddProductPage.picture));
     test('should click on "CREATE A CATEGORY"', () => client.scrollWaitForExistAndClick(AddProductPage.product_create_category_btn, 50));
     test('should set the "New category name"', () => client.waitAndSetValue(AddProductPage.product_category_name_input, data.virtual.new_category_name + date_time));
-    test('should click on "Create"', () => client.createCategory());
+    test('should click on "Create" button', () => client.createCategory());
    // test('should open all categories', () => client.openAllCategories());  //TODO: Verify if we should close then open all categories
     test('should choose the created category as default', () => {
       return promise
@@ -49,11 +49,7 @@ scenario('Create virtual Product in the Back Office', client => {
     test('should set the "Reference"', () => client.waitAndSetValue(AddProductPage.product_reference, data.common.product_reference));
     test('should switch the product online', () => {
       return promise
-        .then(() => {
-          if (global.ps_mode_dev) {
-            client.waitForExistAndClick(AddProductPage.symfony_toolbar);
-          }
-        })
+        .then(() => client.waitForSymfonyToolbar(AddProductPage, 2000))
         .then(() => client.waitForExistAndClick(AddProductPage.product_online_toggle, 3000));
     });
   }, 'product/product');
