+++ tests/E2E/test/selectors/BO/advancedParameters/index.js
@@ -1,0 +1,3 @@
+module.exports = Object.assign(
+  require('./performance')
+);
