+++ tests/E2E/test/selectors/BO/advancedParameters/performance.js
@@ -1,0 +1,7 @@
+module.exports = {
+  Performance: {
+    enableDebugMode: '//label[contains(@for, "form_debug_mode_debug_mode_1")]',
+    save_button: '(//button[text()="Save"])[%I]',
+    success_box: '//*[@id="main-div"]//div[contains(@class, "success")]//div[contains(@class, "alert-text")]',
+  }
+};
